requirements.txt
numpy
pandas
pillow
scikit-learn
xgboost
flask